Implement focusQuality functions & brightQuality functions

+++ imageProcessAlgor/brightQuality.h
@@ -19,8 +19,8 @@ extern "C"
                                         int step,
                                         CaptureImage captureImage);
     __declspec(dllexport) int BrightQuality(cv::Mat& image);
-    int StatBrightnessMean(cv::Mat& image); 
-    int StatBrightnessRMS(cv::Mat& image); 
-    int StatBrightnessFormula(cv::Mat& image); 
-    int StatBrightnessRMSFormula(cv::Mat& image); 
+    float StatBrightnessMean(cv::Mat& image); 
+    float StatBrightnessRMS(cv::Mat& image); 
+    float StatBrightnessFormula(cv::Mat& image); 
+    float StatBrightnessRMSFormula(cv::Mat& image); 
 }
